+++ recipes/default.rb
@@ -3,6 +3,6 @@
 # Recipe:: default
 #
 # Copyright (C) 2013 YOUR_COPYRIGHT
-# 
+#
 # All rights reserved - Do Not Redistribute
 #