+++ spec/tester_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 # Historical Note: For those coming across this cookbook, these
-# filenames were of the original author's doing (Seam O'Meara).
+# filenames were of the original author's doing (Sean O'Meara).
 # There is no special meaning behind them, so don't be confused.
 # He's just weird like the rest of us.
 
@@ -24,4 +24,3 @@ describe 'line::tester' do
     expect(chef_run).to render_file('/tmp/serial.conf')
   end
 end
-